requirements.txt
fastapi
uvicorn
gitpython
requests
pytest